requirements.txt
gradio==5.47.2
numpy==1.24.4
pillow==11.3.0
scikit-image==0.25.2
opencv-python==4.9.0.80